Add inline temperature conversions + minor cleanup
Previously the script assumed temperatures would be defined by variable T and accessed via this. Therefore explicit values in place of '${T}' would not be converted. This addresses that issue in order to allow for potentials directly taken from online documentation.

Also includes minor reformatting and cleanup of the code and comments.